requirements/base.txt
Django==1.10.5

# environment handling
django-environ==0.4.0

# Test helpers
fake-factory==0.5.3
factory_boy==2.5.2

coreapi==2.2.4
coreschema==0.0.4
django-authtools==1.5.0
django-cors-headers==2.0.2
django-email-log==0.2.0
django-extensions==1.7.6
django-filter==1.0.1
django-mail-templated==2.6.5
django-model-utils==2.6.1
django-rest-swagger==2.1.1
djangorestframework==3.5.4
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==0.23
openapi-codec==1.3.1
psycopg2==2.6.2
requests==2.13.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0

